+++ core/modules/pattern.php
@@ -24,7 +24,7 @@ class Pattern {
       'mobile'  => 'icon',
       'content' => null,
       'title'   => isset($args['label']) ? strip_tags($args['label']) : null,
-      'right'   => false
+      'float'   => a::get($this->core->config->element($this->element), 'float', 'left')
     ], $args));
   }