+++ widget/assets/js/widget.js
@@ -1,38 +1,65 @@
 
-var list     = $('.panelBar-widget__list');
+var panelBarList = $('.panelBar-widget__list');
 
 // =============================================
 //  Sortable
 // =============================================
-var sortable = Sortable.create(list[0], {
-  onUpdate: setElements
+var sortable = Sortable.create(panelBarList[0], {
+  onUpdate: setPanelBarElements
 });
 
 // =============================================
 //  On checking/unchecking
 // =============================================
-list.find('input').change(function(e) {
+panelBarList.find('input').change(function(e) {
   var el     = $(this);
-  var active = list.find(':checked').not(el);
+  var active = panelBarList.find(':checked').not(el);
+
+  var floats = el.siblings('.floats').find('a');
+  floats.removeClass('active');
+  if(el.is(':checked')) floats.first().addClass('active');
+
   el.parent().insertAfter(active.last().parent());
-  setElements();
+  setPanelBarElements();
 });
+
+// =============================================
+//  On float setting
+// =============================================
+panelBarList.find('.floats > a').click(function(e) {
+  e.preventDefault();
+  $(this).toggleClass('active');
+  $(this).siblings().removeClass('active');
+  setPanelBarElements();
+  return false;
+});
+
 
 // =============================================
 //  Set API call
 // =============================================
-function setElements() {
-  var data = { 'elements[]' : []};
-  list.find(":checked").each(function() {
-    data['elements[]'].push($(this).val());
+function setPanelBarElements() {
+  var data = generatePanelBarConfig();
+  $.post(setPanelBarURL, data);
+}
+
+function generatePanelBarConfig() {
+  var data = { 'elements' : []};
+
+  panelBarList.find(":checked").each(function() {
+    data.elements.push({
+      element: $(this).val(),
+      float:   $(this).siblings('.floats').find('.active').data('float')
+    });
   });
-  $.post(setURL, data);
+
+  return data;
 }
 
 // =============================================
 //  Reset button
 // =============================================
-list.prev('h2').find('a').click(function(e) {
+panelBarList.prev('h2').find('a').click(function(e) {
   e.preventDefault();
   $.post($(this).attr("href"));
   location.reload();