+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [2.2.0](https://github.com/distantnative/panel-bar/releases/tag/2.2.0) (2017-03-?)
+:sparkles: Set element float via panel widget
 :balloon: Translations now follow the panel user's language  
 :balloon: Translations now get passed {$variables} to be more flexible 
 :poop: Dropped RTL support for the moment